+++ changelog.md
@@ -1,5 +1,6 @@
 # 1.0.0
 - Bump to Geode v5.1.0
+- Add image indicator support
 
 # 1.0.0-beta.1
 - Initial release